// src/types/i18n-types.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface I18N {
  page?: {
    title?: string;
    description?: string;
    privacy?: string;
    legal_notice?: string;
    licenses?: string;
    contact?: string;
    [k: string]: any;
  };
  header?: {
    home?: string;
    about?: string;
    projects?: string;
    contact?: string;
    contact_us?: string;
    [k: string]: any;
  };
  footer?: {
    build_with?: string;
    by_team_and?: string;
    contributers?: string;
    rights_reserved?: string;
    [k: string]: any;
  };
  homepage?: {
    title?: string;
    description?: string;
    core_values?: {
      privacy?: {
        title?: string;
        description?: string;
        [k: string]: any;
      };
      transparency?: {
        title?: string;
        description?: string;
        [k: string]: any;
      };
      security?: {
        title?: string;
        description?: string;
        [k: string]: any;
      };
      tracking?: {
        title?: string;
        description?: string;
        [k: string]: any;
      };
      [k: string]: any;
    };
    [k: string]: any;
  };
  test?: {
    helloworld?: string;
    [k: string]: any;
  };
  aboutPage?: {
    header?: {
      title?: string;
      highlitedTitle?: string;
      content1?: string;
      content2?: string;
      [k: string]: any;
    };
    whoWeAre?: {
      title?: string;
      content?: string;
      [k: string]: any;
    };
    history?: {
      title?: string;
      history1?: {
        title?: string;
        date?: string;
        description?: string;
        [k: string]: any;
      };
      history2?: {
        title?: string;
        date?: string;
        description?: string;
        [k: string]: any;
      };
      history3?: {
        title?: string;
        date?: string;
        description?: string;
        [k: string]: any;
      };
      [k: string]: any;
    };
    card1?: {
      title?: string;
      description?: string;
      [k: string]: any;
    };
    card2?: {
      title?: string;
      description?: string;
      [k: string]: any;
    };
    [k: string]: any;
  };
  [k: string]: any;
}